refactor to avoid circular dependency
